hateoas implemented and start to implement tests

+++ src/main/java/com/framework/blog/service/AlbumService.java
@@ -1,42 +1,56 @@
 package com.framework.blog.service;
 
 import com.framework.blog.exception.AlbumNotExist;
+import com.framework.blog.exception.PermissionDeniedException;
 import com.framework.blog.model.Album;
+import com.framework.blog.model.UserBlog;
 import com.framework.blog.repository.AlbumRepository;
+import com.framework.blog.repository.UserBlogRepository;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
-import javax.validation.Valid;
+import java.util.List;
 
 @Service
 public class AlbumService {
 
     private final AlbumRepository albumRepository;
+    private final UserBlogService userBlogService;
 
-    public AlbumService(AlbumRepository albumRepository) {
+    public AlbumService(AlbumRepository albumRepository, UserBlogService userBlogService) {
         this.albumRepository = albumRepository;
+        this.userBlogService = userBlogService;
     }
 
-    public Album createAlbum(Album album){
-        return albumRepository.save(album);
+    public Album createAlbum(){
+        UserBlog userBlog = userBlogService.userLogged();
+        return albumRepository.save(new Album(userBlog));
     }
 
-    public void deleteAlbum(Long id) throws AlbumNotExist {
-        albumRepository.deleteById(checkAlbumExist(id).getId());
+    public void deleteAlbum(Long id) throws AlbumNotExist, PermissionDeniedException {
+        UserBlog userBlog = userBlogService.userLogged();
+        Album album = albumRepository.getById(id);
+        if (userBlog.getId() == album.getUserBlog().getId()){
+            albumRepository.deleteById(checkAlbumExist(id).getId());
+        }else{
+            throw new PermissionDeniedException("You don't have permission to delete this Album!");
+        }
     }
 
     public Album findById(Long id) throws AlbumNotExist {
         return checkAlbumExist(id);
     }
 
-    public Page<Album> findAllAlbums(Pageable pageable){
-        return albumRepository.findAll(pageable);
+    public List<Album> findAll(){
+        return albumRepository.findAll();
     }
 
-    private Album checkAlbumExist(Long id) throws AlbumNotExist {
+    public Album checkAlbumExist(Long id) throws AlbumNotExist {
         return albumRepository.findById(id)
                 .orElseThrow(()-> new AlbumNotExist("Cannot find album with id: "+ id +" /n" +
                         "Please insert another id and try again!"));
     }
+
 }
